WechatGroupQRCode/requirements.txt
requests>=2.22.0
Pillow>=5.4.1
lxml>=3.2.1
bs4>=0.0.1
beautifulsoup4>=4.7.1
pyzbar>=0.1.8
zbar>=0.10